update: change password and delete account section changed to widgets
