+++ lib/core/constrants/app_fonts.dart
@@ -1,0 +1,3 @@
+class AppFonts {
+  static final poppins='Poppins';
+}
