+++ lib/core/constrants/app_texts.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+import 'package:utracker/core/constrants/app_colors.dart';
+
+var myTextTheme = TextTheme(
+          bodyMedium: const TextStyle(fontSize: 16, color: Colors.black),
+          titleLarge: const TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
+          titleSmall: const TextStyle(fontSize: 14, color: Colors.grey),
+          labelLarge: TextStyle(fontSize: 14, color: AppColors.mainGreenColor),
+        );
