+++ lib/features/habit_tracker/presentation/functions/is_today.dart
@@ -1,0 +1,9 @@
+bool isToday(DateTime? selectedDay) {
+      DateTime today = DateTime.now();
+      today = DateTime(today.year, today.month, today.day);
+      if (today == selectedDay) {
+        return true;
+      } else {
+        return false;
+      }
+    }
